chore: rename keys inline with containers
